requirementsold.txt
beautifulsoup4==4.8.0
bs4==0.0.1
Django==2.2.3
django-bootstrap3==11.1.0
django-bootstrap4==0.0.8
django-datatable-view==0.9.0
django-inline-actions==2.2.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
